+++ CHANGES.rst
@@ -3,6 +3,8 @@ Changelog (nionui)
 
 0.4.0 (UNRELEASED)
 ------------------
+- Add show_ok_dialog to application.
+- Add run_ui to application for running simple declarative windows.
 - Add functions for opening file dialogs without a window.
 - Add support for adjusting menus at the application level.
 - Add function to truncate string to pixel width.